[SDK-72]: Excluded Google.Protobuf assemblies from Code Coverage
Excluded assemblies produced by the Google.Protobuf NuGet package. These are auto-generated, and contain code which we aren't utlising in this project. It doesn't make sense to write code to test the parts which we're not using. For the next check-in I'll look at testing every part of this ProtoBuf that we integrate with.