SDK-1595: Remove new YotiClient initialisation (not released), as this seems to cause a strange bug
This strangely causes a problem when building the .NET 4.7 Example project. When initialising the YotiClient with a DIFFERENT initialiser, the codepath does not go through this new method, but the compiler seems to get confused, and think that a reference is required to BouncyCastle/Portable.BouncyCastle. I think this is a bug with .NET. Using named parameters from integrator's code gets round this, but I think having clients specify "new HttpClient()" if they need to, is more preferable than having them specify named parameters OR adding the portable.BouncyCastle reference.