SDK-1842: Add AuthType to IDV (Doc Scan)
The Identity verification service can accept a value for configuring the
Authentication Scheme to use in HTTP Authorization Headers for
callback messages to Relying Businesses' systems. The authentication
scheme can be set to either Basic or Bearer. This will be included with
any authorization token set in the notification configuration. This can
be used to help secure callback endpoints and so provide some level of
certainty that notifications originated with Yoti systems.